swagger moved into api-docs
